fix accidental folder deletion

+++ evremixes.py
@@ -73,16 +73,24 @@ if os.path.exists(output_folder) and os.listdir(output_folder):
     if remove_files == "\n" or remove_files == "y":
         remove_files = "y"
 
-    if remove_files == "y":
-        for filename in os.listdir(output_base_folder):
-            file_path = os.path.join(output_base_folder, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print(f"Failed to delete {file_path}. Reason: {e}")
+    # Only execute the delete code if output_folder is a subdirectory of output_base_folder,
+    # not output_base_folder itself
+    if os.path.commonpath([output_folder, output_base_folder]) == output_base_folder and \
+        output_folder != output_base_folder:
+
+        if remove_files == "y":
+            for filename in os.listdir(output_folder):
+                file_path = os.path.join(output_folder, filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except Exception as e:
+                    print(f"Failed to delete {file_path}. Reason: {e}")
+
+    else:
+        print("Deletion aborted: Target folder is not a subfolder of the base directory.")
 
 if not os.path.exists(output_folder):
     os.makedirs(output_folder)
